chore: rename builders -> legacybuilders
