+++ internal/controllers/exec_access_request_controller.go
@@ -28,9 +28,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/diranged/oz/internal/api/v1alpha1"
-	"github.com/diranged/oz/internal/builders"
 	"github.com/diranged/oz/internal/controllers/internal/status"
 	"github.com/diranged/oz/internal/controllers/internal/utils"
+	"github.com/diranged/oz/internal/legacybuilder"
 )
 
 // ExecAccessRequestReconciler reconciles a ExecAccessRequest object
@@ -103,8 +103,8 @@ func (r *ExecAccessRequestReconciler) Reconcile(
 	}
 
 	// Create an AccessBuilder resource for this particular template, which we'll use to then verify the resource.
-	builder := &builders.ExecAccessBuilder{
-		BaseBuilder: builders.BaseBuilder{
+	builder := &legacybuilder.ExecAccessBuilder{
+		BaseBuilder: legacybuilder.BaseBuilder{
 			Client:    r.Client,
 			Ctx:       ctx,
 			APIReader: r.APIReader,
