+++ internal/legacybuilder/base_builder_test.go
@@ -1,4 +1,4 @@
-package builders
+package legacybuilder
 
 import (
 	"context"
