+++ internal/legacybuilder/interfaces.go
@@ -1,4 +1,4 @@
-package builders
+package legacybuilder
 
 import (
 	"context"
