+++ internal/legacybuilder/pod_access_builder.go
@@ -1,4 +1,4 @@
-package builders
+package legacybuilder
 
 import (
 	"errors"
