+++ internal/legacybuilder/suite_test.go
@@ -1,4 +1,4 @@
-package builders
+package legacybuilder
 
 import (
 	"path/filepath"
